Moved static formatTime() from Game to utils.
