+++ templates/achievement-stat.mako
@@ -1,4 +1,7 @@
 <%page args="ach, games"/>
+<%!
+import tntfl.templateUtils as utils
+%>
 <div class="col-sm-3">
   <div class="panel panel-default panel-statbox panel-achievement" title="${ach.description}">
     <div class="panel-heading">
@@ -13,7 +16,7 @@
       <ul class="list-unstyled achievement-games" id="achievement-${ach.__name__}-collapse">
 %for game in games:
         <li>
-            <a href="../../game/${game.time}">${game.formatTime(game.time)}</a>
+            <a href="../../game/${game.time}">${utils.formatTime(game.time)}</a>
         </li>
 %endfor
       </ul>
