+++ tntfl/templateUtils.py
@@ -1,3 +1,4 @@
+from datetime import date, datetime, timedelta
 
 def getNumYellowStripes(player, games):
     return len([g for g in games if (g.redPlayer == player.name and g.redScore == 10 and g.blueScore == 0) or (g.bluePlayer == player.name and g.blueScore == 10 and g.redScore == 0)])
@@ -9,4 +10,12 @@ def punditryAvailable(pundit, game, ladder):
     red = ladder.getPlayer(game.redPlayer)
     blue = ladder.getPlayer(game.bluePlayer)
     return pundit.anyComment(red, game, blue)
-    
+
+def formatTime(inTime):
+    time = datetime.fromtimestamp(float(inTime))
+    dateStr = time.strftime("%Y-%m-%d %H:%M")
+    if date.fromtimestamp(float(inTime)) == date.today():
+        dateStr = "%02d:%02d" % (time.hour, time.minute)
+    elif date.fromtimestamp(float(inTime)) > (date.today() - timedelta(7)):
+        dateStr = "%s %02d:%02d" % (("Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun")[time.weekday()], time.hour, time.minute)
+    return dateStr
